New high performance attributes implement for cppast.net (#80)
* Enhance for function template & CppType FullName support.

1. add support for function template
2. add inline namespace support
3. change FullName as a CppType property, and deduce the CppClass full name with specialized template right, you can just use CppClass full name in a generated c++ codes now.
4. FindByFullName() now can search auto ignore inline namespace now(such as clang std::__1::vector, now you can just use std::vector to search)
5. fix crash when the CppClass has a  specialized template with PartialSpecializedTemplateDecl
6. fix crash when typedef with a AliasTemplateDecl for Underlying type.

* 1. Add a test for partial specialized template && 2. Add a IsSpecializedArgument for CppTemplateArgument, so we can detect is a specialized argument or not.

* --other=1. move Tokenizer to CppTokenUtil class  && 2. add a sperate list for token parser attributes and mark it as deprecated

* --other= 1. add a document for new attribute,  in location doc/attributes.md   && 2. new attribute support by __cppast() && 3. default system attribute support not use token parser

---------
