+++ src/CppAst/CppGlobalDeclarationContainer.cs
@@ -30,7 +30,8 @@ namespace CppAst
             Classes = new CppContainerList<CppClass>(this);
             Typedefs = new CppContainerList<CppTypedef>(this);
             Namespaces = new CppContainerList<CppNamespace>(this);
-            Attributes = new CppContainerList<CppAttribute>(this);
+            Attributes = new List<CppAttribute>();
+            TokenAttributes = new List<CppAttribute>();
         }
 
         /// <summary>
@@ -60,7 +61,10 @@ namespace CppAst
         public CppContainerList<CppNamespace> Namespaces { get; }
 
         /// <inheritdoc />
-        public CppContainerList<CppAttribute> Attributes { get; }
+        public List<CppAttribute> Attributes { get; }
+
+        [Obsolete("TokenAttributes is deprecated. please use system attributes and annotate attributes")]
+        public List<CppAttribute> TokenAttributes { get; }
 
         /// <inheritdoc />
         public virtual IEnumerable<ICppDeclaration> Children()
