+++ src/CppAst/CppNamespace.cs
@@ -25,7 +25,8 @@ namespace CppAst
             Classes = new CppContainerList<CppClass>(this);
             Typedefs = new CppContainerList<CppTypedef>(this);
             Namespaces = new CppContainerList<CppNamespace>(this);
-            Attributes = new CppContainerList<CppAttribute>(this);
+            Attributes = new List<CppAttribute>();
+            TokenAttributes = new List<CppAttribute>();
         }
 
         /// <summary>
@@ -57,7 +58,10 @@ namespace CppAst
         public CppContainerList<CppNamespace> Namespaces { get; }
 
         /// <inheritdoc />
-        public CppContainerList<CppAttribute> Attributes { get; }
+        public List<CppAttribute> Attributes { get; }
+        [Obsolete("TokenAttributes is deprecated. please use system attributes and annotate attributes")]
+        public List<CppAttribute> TokenAttributes { get; }
+
 
         protected bool Equals(CppNamespace other)
         {
